fix,solaris: call of overloaded 'log(size_t&)' is ambiguous
   return depCost + nrItems * log(nrItems);
                                         ^
arangod/Aql/SortNode.cpp:196:41: note: candidates are:
In file included from /opt/local/gcc49/lib/gcc/x86_64-sun-solaris2.11/4.9.3/include-fixed/math.h:47:0,
                 from ./lib/Basics/Common.h:62,
                 from arangod/Aql/SortNode.h:27,
                 from arangod/Aql/SortNode.cpp:24:
/opt/local/gcc49/lib/gcc/x86_64-sun-solaris2.11/4.9.3/include-fixed/iso/math_iso.h:221:21: note: long double std::log(long double)
  inline long double log(long double __X) { return __logl(__X); }
                     ^
/opt/local/gcc49/lib/gcc/x86_64-sun-solaris2.11/4.9.3/include-fixed/iso/math_iso.h:189:15: note: float std::log(float)
  inline float log(float __X) { return __logf(__X); }
               ^
/opt/local/gcc49/lib/gcc/x86_64-sun-solaris2.11/4.9.3/include-fixed/iso/math_iso.h:89:15: note: double std::log(double)
 extern double log __P((double));
               ^

+++ arangod/Aql/SortNode.cpp
@@ -193,6 +193,6 @@ double SortNode::estimateCost(size_t& nrItems) const {
   if (nrItems <= 3.0) {
     return depCost + nrItems;
   }
-  return depCost + nrItems * log(nrItems);
+  return depCost + nrItems * log(static_cast<double>(nrItems));
 }
